requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
greenlet==3.1.1
idna==3.10
numpy==1.26.3
pandas==2.2.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
